requirements.txt
fastapi
uvicorn
celery
redis
minio
pydantic
Pillow
python-multipart
python-dotenv